Refactoring: - renamed class to more appropriate - added read/write locks to make thread safe - moved file injection to constructor parameter
